Use the search method and paging to find Gtilab project match
